체력,산소,데미지 텍스트 통합
TextPopup 프리팹으로 통합, 코드도 TextPopup.cs 로 통합.
앞으로 체력관련 텍스트 생서할때 Set~~ 이용해서 값만 넣어주면 색상은 알아서 들어감
