--- Assets/03.Prefabs/DamageCanvas.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: cc5a5cf014803be4394a862a82ee9032
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
