--- Assets/03.Prefabs/DamageObject.prefab
@@ -1,172 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!1 &3831522259552869590
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 1237823570391400854}
-  - component: {fileID: 4284185697113995631}
-  m_Layer: 0
-  m_Name: DamageObject
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &1237823570391400854
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3831522259552869590}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 3.55, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children:
-  - {fileID: 1905585142522216518}
-  m_Father: {fileID: 0}
-  m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &4284185697113995631
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3831522259552869590}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a83973076092e384c9d3f54000465452, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!1001 &3244886832957022541
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 1237823570391400854}
-    m_Modifications:
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_Pivot.x
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_Pivot.y
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_AnchorMax.x
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_AnchorMax.y
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_AnchorMin.x
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_AnchorMin.y
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_SizeDelta.x
-      value: 2
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_SizeDelta.y
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_AnchoredPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_AnchoredPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 180
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3997464698092151095, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-        type: 3}
-      propertyPath: m_Name
-      value: DamageText
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 1321e8bf32ccb1344811fec23cd2dd5d, type: 3}
---- !u!224 &1905585142522216518 stripped
-RectTransform:
-  m_CorrespondingSourceObject: {fileID: 3997464698092151051, guid: 1321e8bf32ccb1344811fec23cd2dd5d,
-    type: 3}
-  m_PrefabInstance: {fileID: 3244886832957022541}
-  m_PrefabAsset: {fileID: 0}
