--- Assets/03.Prefabs/DamageObject.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: fbb7fba10039235498acd5225bca83cc
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
