+++ Assets/03.Prefabs/TextPopup.prefab
@@ -12,8 +12,9 @@ GameObject:
   - component: {fileID: 3997464698092151050}
   - component: {fileID: 3997464698092151049}
   - component: {fileID: 3997464698092151048}
+  - component: {fileID: 7069438660160322206}
   m_Layer: 0
-  m_Name: DamageText
+  m_Name: TextPopup
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -131,7 +132,7 @@ MonoBehaviour:
   m_fontColor32:
     serializedVersion: 2
     rgba: 4278190335
-  m_fontColor: {r: 1, g: 0, b: 0, a: 1}
+  m_fontColor: {r: 1, g: 1, b: 1, a: 1}
   m_enableVertexGradient: 0
   m_colorMode: 3
   m_fontColorGradient:
@@ -193,3 +194,15 @@ MonoBehaviour:
   m_hasFontAssetChanged: 0
   m_renderer: {fileID: 3997464698092151049}
   m_maskType: 0
+--- !u!114 &7069438660160322206
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3997464698092151095}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a83973076092e384c9d3f54000465452, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
