+++ Assets/GamVal/PlayerTakeItemTest/PlayerTakeItem.cs
@@ -6,6 +6,7 @@ public class PlayerTakeItem : MonoBehaviour
 {
     PlayerInfo playerInfo;
     public HealthBar healthBar;
+    public TextPopup TextPopup;
     [SerializeField]
     float healAmount = 300.0f;
     [SerializeField]
@@ -20,6 +21,8 @@ public class PlayerTakeItem : MonoBehaviour
     {
         if (other.CompareTag("HEALITEM"))
         {
+            TextPopup textPopup = Instantiate<TextPopup>(TextPopup, other.transform.position, other.transform.rotation);
+            textPopup.GetComponent<TextPopup>().SetHealAmount((int)healAmount);
             playerInfo.hp += healAmount;
             if (playerInfo.hp > 1000.0f)
             {
@@ -31,6 +34,8 @@ public class PlayerTakeItem : MonoBehaviour
         }
         else if (other.CompareTag("OXYGENITEM"))
         {
+            TextPopup textPopup = Instantiate<TextPopup>(TextPopup, other.transform.position, other.transform.rotation);
+            textPopup.GetComponent<TextPopup>().SetOxygenAmount((int)oxygenAmount);
             playerInfo.oxygen += oxygenAmount;
             if (playerInfo.oxygen > 1000.0f)
             {
